Refactor x509 parsing code into module separate from yubikey
